feat: Complete project-wide config refactor and bug fixes
