+++ js/config/index.js
